Small improvement to getChromInfoFromNCBI()
Before this change getChromInfoFromNCBI() was calling read.table()
directly on the URL of the NCBI assembly report. Unfortunately
calling read.table() on an URL tends to be unreliable. For example
getChromInfoFromNCBI("CIEA01") was randomly failing for me with a
"line 18564 did not have 10 elements" error (exact line number would
vary at each new attempt).

Now getChromInfoFromNCBI() proceeds in 2 steps: first it downloads the
file, then it calls read.table() on the local file. This seems to be a
lot more reliable!